Move umount_mdv method into infra module
That is the only place it seems to be used, so it might as well go there.

Note that it will break when devlinks are moved from "/stratis" to
"/dev/stratis".
